move haversine to source code
1. remove precompiled jar for f-droid compliance
2. implement haversine calculation according to wikipedia
