+++ androidWeather/src/main/java/ca/fwe/weather/util/LocationFetcher.java
@@ -5,7 +5,7 @@ import android.location.Location;
 import android.location.LocationListener;
 import android.location.LocationManager;
 import android.os.Bundle;
-import ca.fwe.locations.geometry.LatLon;
+import ca.fwe.weather.util.LatLon;
 
 public class LocationFetcher {
